backend/predict/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
onnxruntime==1.23.2          # <- use this (available per your pip error list)
pydantic==2.9.2
python-dotenv==1.0.1
pymongo==4.9.1
numpy==1.26.4
